feat: add make to src
